Prefix the moved scripts to old_ to avoid confusion
This should prevent some confusion for now whilst we decide how to handle the dependency in the process check links script
